<ctype.h>: more optimization.
clang is pretty good at turning something like

    (c >= '0' && c <= '9')

into the equivalent of an array bounds check:

    ((unsigned c)-'0' < 10).

Unfortunately, it struggles if you combine two or more of these
comparisons.

So although it generates [afaik] optimal code for isdigit(), and can
separately do the same for

    (c >= 'a' && c <= 'f'),

or

    (c >= 'A' && c <= 'F'),

or even

    (_tolower(c) >= 'a' && _tolower(c) <= 'f'),

when you combine them (as in isxdigit()) it no longer recognizes any of
these as something it can optimize and puts all the branches back in.

To avoid this, let's add a macro that does the optimized range check and
use that throughout <ctype.h>.

With the new more accurate <ctype.h> benchmarks, we see improvements
of around 40-50MiB/s more for arm64 (not counting isspace(), which
skews the statistics by being roughly twice as fast, going from
216MiB/s to 438MiB/s). arm32 benefits less, from "not at all" for
isprint() or isspace(), to 40MiB/s better for islower().

Only ispunct() still has an explicit branch for arm64; all the arm32
functions lose their branches, if you don't count the `bxlo lr` in
the same part of ispunct().

Test: treehugger
Change-Id: I663ff7e31b0f4173d6dd27f16bb2d8d3485c8e06