sigqueue/pthread_sigqueue: cleanup.
Remove the explicit memset() from one of the two callers, and use explicit initialization for both. This makes it clearer that optimizing the zeroing of the rest of the struct is the compiler's problem, not ours.

Add the "missing" assignment to si_signo in pthread_sigqueue() that isn't actually necessary because the kernel will overwrite that field anyway when it copies from userspace, but which looked like a bug given the difference between the sigqueue() and pthread_sigqueue() implementations.

Also reuse the result of getpid() in pthread_sigqueue() rather than calling it twice.

Change-Id: I39578d80ddc5edcb7d235b078392e20f35713dba